only me

+++ index.js
@@ -10,9 +10,11 @@ client.on('ready', () => {
 client.on('message', msg => {
   if (!msg.author.bot) {
     if (msg.content.startsWith(config.prefix + "say ")) {
+      if (msg.author.id == "432380402238357516"){
       var say = msg.content.replace(config.prefix + "say ","");
       msg.delete();
-      msg.channel.send(say);
+      msg.channel.send(say);}
+      else {msg.reply("no")}
     }
     if (msg.channel.id === config.reactchan) {
   if (msg.content.startsWith(config.prefix)) {
